+++ R/abbyyR.R
@@ -19,7 +19,7 @@
 #' @importFrom httr GET POST authenticate config stop_for_status upload_file content
 #' @importFrom XML  xmlToList
 #' @importFrom curl curl_download curl_fetch_memory
-#' @importFrom RecordLinkage levenshteinDist
+#' @importFrom progress progress_bar
 #' @importFrom readr read_file
 #' @docType package
 #' @author Gaurav Sood